requirements.txt
flask
flask_sqlalchemy
pymysql
cryptography
SQLAlchemy
bcrypt
pyjwt
flask_cors
psycopg2-binary
flask_migrate 
python-dotenv
